requirements.txt
Django==4.0.10
PyJWT==2.10.1
PyYAML==6.0.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset_normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj_rest_auth==2.1.11
django_allauth==0.50.0
django_cors_headers==3.10.1
djangorestframework==3.13.1
drf_spectacular==0.21.2
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema_specifications==2024.10.1
oauthlib==3.2.2
pillow==11.1.0
pip==24.0
pycparser==2.22
python3_openid==3.2.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests_oauthlib==2.0.0
rpds_py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
